todo-service/src/test/java/blog/vgarg/todos/resources/TodoControllerTest.java
package blog.vgarg.todos.resources;

import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;

import java.util.List;

import static org.assertj.core.api.Assertions.assertThat;

@SpringBootTest
public class TodoControllerTest {

    @Autowired
    private TodoController todoController;

    @Test
    public void testTodos() {
        ResponseEntity<List<String>> response = todoController.getTodos();
        assertThat(response).isNotNull();
        assertThat(response.getStatusCode()).isEqualTo(HttpStatus.OK);
        assertThat(response.getBody()).hasSize(0);
    }
}
